refactor: reorganize source, errors
