+++ packages/@sanity/cli-core/src/_exports/errors.ts
@@ -1,0 +1,5 @@
+export * from '../errors/CLIError.js'
+export * from '../errors/CLIWarning.js'
+export * from '../errors/NonInteractiveError.js'
+export * from '../errors/NotFoundError.js'
+export * from '../errors/ProjectRootNotFoundError.js'
